safari-600.7-branch/LayoutTests/batterystatus/window-property-expected.txt

Tests that the window.navigator.battery properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.webkitBattery == 'object' is true
PASS hasOnBatteryProperty() is true
PASS navigator.webkitBattery.hasOwnProperty('onchargingchange') is true
PASS navigator.webkitBattery.hasOwnProperty('onchargingtimechange') is true
PASS navigator.webkitBattery.hasOwnProperty('ondischargingtimechange') is true
PASS navigator.webkitBattery.hasOwnProperty('onlevelchange') is true
PASS successfullyParsed is true

TEST COMPLETE
